Refactor : sendReadyHandler의 game관련 로직을 gameController로 이동
이유:
- 해당 handler를 제외하고도 다른 곳에서 위의 로직을 사용하는 경우가 발생
