+++ server/socket/handlers/sendReadyHandler.js
@@ -1,45 +1,6 @@
 const rooms = require('../rooms');
 const io = require('../io');
-
-const assignViewer = socket => {
-  const room = rooms.getRoomByRoomId(socket.roomId);
-  const { streamerSocketId } = room;
-  const socketIds = Object.keys(room.players);
-  socketIds.forEach(socketId => {
-    if (socketId !== streamerSocketId) {
-      io.to(socketId).emit('assignViewer', { streamerSocketId });
-    }
-  });
-};
-
-const assignStreamer = socket => {
-  const { streamerSocketId } = rooms.getRoomByRoomId(socket.roomId);
-  io.to(streamerSocketId).emit('assignStreamer');
-};
-
-const startRound = roomId => {
-  rooms.setRound(roomId);
-  io.to(roomId).emit('roundStart', {
-    currentRound: rooms.getRoomByRoomId(roomId).currentRound,
-  });
-};
-
-const startSet = roomId => {
-  rooms.setSet(roomId);
-  io.to(roomId).emit('setStart', {
-    currentSet: rooms.getRoomByRoomId(roomId).currentSet,
-  });
-};
-
-const startGame = (socket, roomId) => {
-  rooms.resetGameProgress(roomId);
-  rooms.setRoomStatusByRoomId(roomId, 'playing');
-  io.to(roomId).emit('gameStart');
-  startRound(roomId);
-  startSet(roomId);
-  assignStreamer(socket);
-  assignViewer(socket);
-};
+const { startGame } = require('../gameController');
 
 const setPlayerReady = (socket, isReady) => {
   const room = rooms.findRoomBySocket(socket);
